Initial Commit: add schedule to debt
