+++ db/migrate/20160713211237_add_schedule_to_debts.rb
@@ -1,0 +1,5 @@
+class AddScheduleToDebts < ActiveRecord::Migration
+  def change
+    add_reference :debts, :schedule, index: true, foreign_key: true
+  end
+end
